'feature-search-speed-history-progressbarupload-converttomp3-changeicon'

+++ app/src/main/java/com/example/duan1_customer/model/Result_UploadAudio.java
@@ -3,5 +3,5 @@ package com.example.duan1_customer.model;
 public class Result_UploadAudio {
     public String name;
     public String url;
-    public float minutes;
+    public double minutes;
 }
